+++ src/main/java/erp/greenagro/greenagro_erp_backend/dto/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package erp.greenagro.greenagro_erp_backend.dto.login;
+package erp.greenagro.greenagro_erp_backend.dto.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
