+++ src/main/java/erp/greenagro/greenagro_erp_backend/dto/auth/LoginResponse.java
@@ -1,13 +1,12 @@
-package erp.greenagro.greenagro_erp_backend.dto.login;
+package erp.greenagro.greenagro_erp_backend.dto.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class TokenBundle {
+public class LoginResponse {
 
     private String accessToken;
-    private String refreshToken;
 
 }
