+++ src/main/java/erp/greenagro/greenagro_erp_backend/dto/auth/TokenBundle.java
@@ -1,12 +1,13 @@
-package erp.greenagro.greenagro_erp_backend.dto.login;
+package erp.greenagro.greenagro_erp_backend.dto.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class LoginResponse {
+public class TokenBundle {
 
     private String accessToken;
+    private String refreshToken;
 
 }
